Fix all failing tests in PostgresRepo implementation
- Fixed TestDatabase.swift: Corrected INSERT statement to include password field and proper values alignment
- Fixed UserSchema.swift: Changed score field type from .integer to .float to match DOUBLE PRECISION in database
- Fixed PostgresRepo.swift: Ensured ID field is always included in SELECT queries to prevent "Missing or invalid ID" errors

All tests now pass successfully (27 tests, 0 failures).

+++ Sources/Spectro/SchemaBuilder/UserSchema.swift
@@ -14,7 +14,7 @@ struct UserSchema: Schema {
         Field.description("age", .integer(defaultValue: 0))
         Field.description("password", .string, isRedacted: true)
         Field.description("email", .string)
-        Field.description("score", .integer(defaultValue: 0))
+        Field.description("score", .float(defaultValue: 0))
         Field.description("is_active", .boolean(defaultValue: true))
         Field.description("created_at", .timestamp)
         Field.description("updated_at", .timestamp)
